+++ code/inetfile/chttpget.cpp
@@ -217,7 +217,7 @@ void ChttpGet::WorkerThread()
 		return;
 	}
 	sprintf(szCommand,"GET %s%s HTTP/1.1\nAccept: */*\nAccept-Encoding: deflate\nHost: %s\n\n\n",m_ProxyEnabled?"":"/",m_ProxyEnabled?m_URL:m_szDir,m_szHost);
-	send(m_DataSock,szCommand,strlen(szCommand),0);
+	send(m_DataSock,szCommand,static_cast<int>(strlen(szCommand)),0);
 	p = GetHTTPLine();
 	if (!p) return;
 	if(strnicmp("HTTP/",p,5)==0)
